batch/pyproject.toml
[project]
name = "batch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elasticsearch>=8.0.0,<9.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
